Infra: Rename and refactor canonical ref directive to inject reST links
